+++ appinfo/routes.php
@@ -115,7 +115,12 @@ return [
 		['name' => 'OCSUi#unregisterExAppMenuEntry', 'url' => '/api/v1/ui/top-menu', 'verb' => 'DELETE'],
 		['name' => 'OCSUi#getExAppMenuEntry', 'url' => '/api/v1/ui/top-menu', 'verb' => 'GET'],
 
-		//Common UI
+		// ExApp routes
+		['name' => 'ExAppRoutes#registerExAppRoutes', 'url' => '/api/v1/routes', 'verb' => 'POST'],
+		['name' => 'ExAppRoutes#unregisterExAppRoutes', 'url' => '/api/v1/routes', 'verb' => 'DELETE'],
+		['name' => 'ExAppRoutes#getExAppRoutes', 'url' => '/api/v1/ui/routes', 'verb' => 'GET'],
+
+		// Common UI
 		['name' => 'OCSUi#setExAppInitialState', 'url' => '/api/v1/ui/initial-state', 'verb' => 'POST'],
 		['name' => 'OCSUi#deleteExAppInitialState', 'url' => '/api/v1/ui/initial-state', 'verb' => 'DELETE'],
 		['name' => 'OCSUi#getExAppInitialState', 'url' => '/api/v1/ui/initial-state', 'verb' => 'GET'],